Move Ember.DefaultResolver to @ember/application/globals-resolver.
