+++ packages/@ember/application/globals-resolver.js
@@ -6,7 +6,7 @@ import { dictionary } from 'ember-utils';
 import { get, findNamespace } from 'ember-metal';
 import { assert, info } from 'ember-debug';
 import { String as StringUtils, Object as EmberObject } from 'ember-runtime';
-import validateType from '../utils/validate-type';
+import validateType from './lib/validate-type';
 import { getTemplate } from 'ember-glimmer';
 import { DEBUG } from 'ember-env-flags';
