+++ packages/@ember/application/lib/validate-type.js
