+++ packages/@ember/application/tests/dependency_injection/default_resolver_test.js
