+++ packages/@ember/application/tests/dependency_injection/normalization_test.js
