+++ packages/ember-application/index.js
@@ -1,5 +1,4 @@
 export { default as ApplicationInstance } from './lib/system/application-instance';
-export { default as Resolver } from './lib/system/resolver';
 export { default as Engine } from './lib/system/engine';
 export { default as EngineInstance } from './lib/system/engine-instance';
 export { getEngineParent, setEngineParent } from './lib/system/engine-parent';
