+++ packages/internal-test-helpers/lib/test-cases/default-resolver-application.js
@@ -1,5 +1,5 @@
 import AbstractApplicationTestCase from './abstract-application';
-import { Resolver as DefaultResolver } from 'ember-application';
+import DefaultResolver from '@ember/application/globals-resolver';
 import Application from '@ember/application';
 import { setTemplates, setTemplate } from 'ember-glimmer';
 import { assign } from 'ember-utils';
